Add files via upload
This is an HTA file that will show you computers that are active on your network.  It uses the computer names, will ping the computer once every minute (and that is adjustable) and will show those active as black and those not responding as red (in default code) also shows how many times the computer has "dropped off" the network. Feel free to use this code and modify it as needed.  I never found exactly what I was looking for so this is what I created and it worked for our needs like a charm.